Mostly got skyboxes working.  Still one graphical bug

+++ GameEngine/Engine/Events.h
@@ -64,7 +64,7 @@ namespace Events
 	//*************MouseClickEvents***************
 	struct MousePressedEvent : Event
 	{
-		float mouseCode;
+		int mouseCode;
 
 		MousePressedEvent(float keyCode) :
 			mouseCode(keyCode){}
